refactor: Derive UartStm flavours (DMA, DuplexDMA) from base class
