+++ hal_st/stm32fxxx/UartStmDuplexDma.hpp
@@ -2,14 +2,14 @@
 #define HAL_UART_STM_DUPLEX_DMA_HPP
 
 #include "hal_st/stm32fxxx/DmaStm.hpp"
-#include "hal_st/stm32fxxx/UartStm.hpp"
+#include "hal_st/stm32fxxx/UartStmDma.hpp"
 #include <atomic>
 #include <cstdint>
 
 namespace hal
 {
     class UartStmDuplexDma
-        : public UartStm
+        : public UartStmDma
     {
     public:
         using Config = detail::UartStmConfig;
@@ -19,10 +19,6 @@ namespace hal
 
         UartStmDuplexDma(infra::MemoryRange<uint8_t> rxBuffer, hal::DmaStm::TransmitStream& transmitStream, hal::DmaStm::ReceiveStream& receiveStream, uint8_t oneBasedIndex, GpioPinStm& uartTx, GpioPinStm& uartRx, const Config& config = Config());
         UartStmDuplexDma(infra::MemoryRange<uint8_t> rxBuffer, hal::DmaStm::TransmitStream& transmitStream, hal::DmaStm::ReceiveStream& receiveStream, uint8_t oneBasedIndex, GpioPinStm& uartTx, GpioPinStm& uartRx, GpioPinStm& uartRts, GpioPinStm& uartCts, const Config& config = Config());
-#if defined(HAS_PERIPHERAL_LPUART)
-        UartStmDuplexDma(infra::MemoryRange<uint8_t> rxBuffer, hal::DmaStm::TransmitStream& transmitStream, hal::DmaStm::ReceiveStream& receiveStream, uint8_t oneBasedIndex, GpioPinStm& uartTx, GpioPinStm& uartRx, LpUart lpUart, const Config& config = Config());
-        UartStmDuplexDma(infra::MemoryRange<uint8_t> rxBuffer, hal::DmaStm::TransmitStream& transmitStream, hal::DmaStm::ReceiveStream& receiveStream, uint8_t oneBasedIndex, GpioPinStm& uartTx, GpioPinStm& uartRx, GpioPinStm& uartRts, GpioPinStm& uartCts, LpUart lpUart, const Config& config = Config());
-#endif
 
     private:
         UartStmDuplexDma(infra::MemoryRange<uint8_t> rxBuffer, hal::DmaStm::TransmitStream& transmitStream, hal::DmaStm::ReceiveStream& receiveStream, uint8_t oneBasedIndex, GpioPinStm& uartTx, GpioPinStm& uartRx, GpioPinStm& uartRts, GpioPinStm& uartCts, const Config& config, bool hasFlowControl);
@@ -31,26 +27,20 @@ namespace hal
         ~UartStmDuplexDma();
 
         // Implementation of SerialCommunication
-        void SendData(infra::MemoryRange<const uint8_t> data, infra::Function<void()> actionOnCompletion = infra::emptyFunction) override;
         void ReceiveData(infra::Function<void(infra::ConstByteRange data)> dataReceived) override;
 
     private:
         void HalfReceiveComplete();
         void FullReceiveComplete();
         void ReceiveComplete(size_t currentPosition);
-        void TransferComplete();
 
         // Implementation InterruptHandler
         void Invoke() override;
 
     private:
         infra::MemoryRange<uint8_t> rxBuffer;
-        hal::TransmitDmaChannel transmitDmaChannel;
         hal::CircularReceiveDmaChannel receiveDmaChannel;
-        infra::Function<void()> transferDataComplete;
-        infra::Function<void(infra::ConstByteRange data)> dataReceived;
         std::atomic<size_t> lastReceivedPosition{};
-        uint8_t uartIndex;
     };
 }
